+++ Runtime/Scripts/PrimitiveCreateContext.cs
@@ -56,6 +56,7 @@ namespace GLTFast {
 
             Profiler.BeginSample("SetIndices");
             int indexCount = 0;
+            var allBounds = vertexData.bounds;
             for (int i = 0; i < indices.Length; i++) {
                 indexCount += indices[i].Length;
             }
@@ -77,6 +78,14 @@ namespace GLTFast {
                     firstVertex = 0,
                     vertexCount = vertexData.vertexCount
                 };
+                if (allBounds.HasValue) {
+                    // Setting the submeshes' bounds to the overall bounds
+                    // Calculating the actual sub-mesh bounds (by iterating the verts referenced
+                    // by the sub-mesh indices) would be slow. Also, hardly any glTFs re-use
+                    // the same vertex buffer across primitives of a node (which is the
+                    // only way a mesh can have sub-meshes)
+                    subMeshDescriptor.bounds = allBounds.Value;
+                }
                 msh.SetSubMesh(i,subMeshDescriptor,defaultMeshUpdateFlags);
                 Profiler.EndSample();
                 indexCount += indices[i].Length;
@@ -94,9 +103,16 @@ namespace GLTFast {
                 Profiler.EndSample();
             }
             
-            Profiler.BeginSample("RecalculateBounds");
-            msh.RecalculateBounds(); // TODO: make optional! maybe calculate bounds in Job.
-            Profiler.EndSample();
+            if (allBounds.HasValue) {
+                msh.bounds = allBounds.Value;
+            } else {
+                Profiler.BeginSample("RecalculateBounds");
+#if DEBUG
+                Debug.LogError("Bounds have to be recalculated (slow operation). Check if position accessors have proper min/max values");
+#endif
+                msh.RecalculateBounds();
+                Profiler.EndSample();
+            }
 
 #if GLTFAST_KEEP_MESH_DATA
             Profiler.BeginSample("UploadMeshData");