+++ ChangeLog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unlit alpha blended ShaderGraph variants (thanks [@hybridherbst](https://github.com/hybridherbst) for #144)
 - Support for unsigned byte joint indices
 ### Changed
+- Accelerated loading meshes by obtaining and setting bounds from accessors min/max values instead of recalculating them  
 - Improved log message when DracoUnity/KtxUnity packages are missing
 ### Fixed
 - Texture offset/tiling values don't get lost when switching shaders (thanks [@hybridherbst](https://github.com/hybridherbst) for #140)